requirements.txt
rich
art2text
requests[socks]
dnspython
python-whois
beautifulsoup4
PySocks
cryptography
scapy
python-nmap
netifaces
python-wappalyzer
sslyze
certifi
urllib3
aiohttp
asyncio
python-dotenv
fake-useragent
Pillow
piexif
opencv-python
numpy
geopy
stegano
imagehash 